Try to get the condition names right and fix add pruning

+++ genotyping/make-controls.sh
@@ -12,11 +12,15 @@ HG38=../haps/hg38.fa
 chroms=chr21
 #HG38=../haps/hg38_chr21.fa 
 
-#Make a primary control graph and its indexes
-rm -rf jsc ; toil-vg construct ./jsc ./controls --fasta ${HG38} --region ${chroms} --container None --realTimeLogging  --xg_index --gcsa_index --out_name hg38  --primary  --workDir . --gcsa_index_cores 20
+if [[ ! -e ./controls/primary.gcsa.lcp ]] ; then
+    #Make a primary control graph and its indexes
+    rm -rf jsc ; toil-vg construct ./jsc ./controls --fasta ${HG38} --region ${chroms} --container None --realTimeLogging  --xg_index --gcsa_index --out_name hg38  --primary  --workDir . --gcsa_index_cores 20
+fi
 
-#Make a positive control graph and its indexes
-rm -rf jsc ; toil-vg construct ./jsc ./controls --fasta ${HG38} --vcf ../haps/HGSVC.HG00514.vcf.gz --pos_control HG00514 --region ${chroms} --container None --realTimeLogging  --xg_index --gcsa_index --out_name hgsvc.norm --flat_alts --normalize  --workDir . --gcsa_index_cores 20 --whole_genome_config --gbwt_index --gbwt_prune
+if [[ ! -e ./controls/hgsvc.norm_HG00514.gcsa.lcp ]] ; then
+    #Make a positive control graph and its indexes
+    rm -rf jsc ; toil-vg construct ./jsc ./controls --fasta ${HG38} --vcf ../haps/HGSVC.HG00514.vcf.gz --pos_control HG00514 --region ${chroms} --container None --realTimeLogging  --xg_index --gcsa_index --out_name hgsvc.norm --flat_alts --normalize  --workDir . --gcsa_index_cores 20 --whole_genome_config --gbwt_index --gbwt_prune
+fi
 
 
 if [[ ! -e ./controls/hg38.fa.bwt ]] ; then
